# Allow signals to interrupt S_poll for faster shutdown.