This is first pass at making shutdown faster to help with deploy speed. This doesn't actually do a ton in terms of the length of time it takes to shut down a production server, or any non-idle server, but it makes a huge difference in the time it takes to shut down an idle server, which makes a gigantic impact on the time it takes tests to run, as they start and stop servers hundreds of times.

What this change actually does is allow our `poll` loops to be interrupted when a signal is caught. There are three separate places that we wait for up to one second in `poll`, one in `main` waiting on network activity from clients, one in the `sync` thread waiting on network activity from peers, and one in `worker` threads waiting for new commands to become available.

This change adds the ability for `poll` to be interrupted by signals that we use to shut down, and uses this new functionality for the two `poll` loops, so that they return immediately once a signal is caught.

This does not fundamentally change the way the worker threads wait for commands, but does shorten the timeout from 1s to 50ms, which makes them return much more quickly. This results in slightly higher idle CPU usage for these threads, which will affect VMs a bit, but I don't think enough to worry about. The correct fix for this that's in-line with #bedrock3 is to switch to one thread per command, in which case, there will be no idle threads when there are no commands.

Here's performance numbers for this:

Without this change, running a throwaway `FastShutdownTest`:
```
 clustertest:$  ./clustertest -only FastShutdown -repeatCount 10
[--------------]
[ RUN          ] FastShutdownTest::test
Stopped 1 follower in 4003ms, remaining servers in: 8009ms.
Stopped 1 follower in 4001ms, remaining servers in: 7010ms.
Stopped 1 follower in 4017ms, remaining servers in: 7004ms.
Stopped 1 follower in 4006ms, remaining servers in: 6011ms.
Stopped 1 follower in 4002ms, remaining servers in: 7002ms.
Stopped 1 follower in 4009ms, remaining servers in: 6000ms.
Stopped 1 follower in 2000ms, remaining servers in: 6003ms.
Stopped 1 follower in 4003ms, remaining servers in: 7009ms.
Stopped 1 follower in 4004ms, remaining servers in: 7001ms.
Stopped 1 follower in 3002ms, remaining servers in: 4000ms.
[ TEST RESULTS ] Passed: 10, Failed: 0
```

And with this change (at ef0dbd7dda149e5a0e0cf1b91242f6d68d0b2c4f):
```
 clustertest:$  ./clustertest -only FastShutdown -repeatCount 10
[--------------]
[ RUN          ] FastShutdownTest::test
Stopped 1 follower in 39ms, remaining servers in: 98ms.
Stopped 1 follower in 63ms, remaining servers in: 83ms.
Stopped 1 follower in 46ms, remaining servers in: 82ms.
Stopped 1 follower in 15ms, remaining servers in: 100ms.
Stopped 1 follower in 44ms, remaining servers in: 103ms.
Stopped 1 follower in 46ms, remaining servers in: 101ms.
Stopped 1 follower in 52ms, remaining servers in: 98ms.
Stopped 1 follower in 17ms, remaining servers in: 101ms.
Stopped 1 follower in 59ms, remaining servers in: 100ms.
Stopped 1 follower in 53ms, remaining servers in: 102ms.
[ TEST RESULTS ] Passed: 10, Failed: 0
```

And auth tests before the change:
```
test:$  time ./authtest -threads 16 -repeatCount 1
[ TEST RESULTS ] Passed: 1364, Failed: 0
real	4m53.046s
user	1m30.156s
sys	0m49.976s
```

And after the change:
```
test:$  time ./authtest -threads 16 -repeatCount 1
[ TEST RESULTS ] Passed: 1364, Failed: 0
real	2m37.044s
user	2m17.032s
sys	1m36.116s
```